sprint-3/step-1 feat: разбить редьюсер на маленькие редьюсеры
