+++ src/services/actions/burgerConstructorIngredients.js
@@ -1,3 +1,4 @@
 export const ADD_CONSTRUCTOR_ITEM = 'ADD_CONSTRUCTOR_ITEM'
 export const DELETE_CONSTRUCTOR_ITEM = 'DELETE_CONSTRUCTOR_ITEM'
 export const SET_CHOOSEN_BUN = 'SET_CHOOSEN_BUN'
+export const SET_SORTED_ARRAY = 'SET_SORTED_ARRAY'
