+++ src/services/reducers/detailIngredientReducer.js
@@ -1,0 +1,24 @@
+import { CLEAR_DETAIL_INGREDIENT, SET_DETAIL_INGREDIENT } from '../actions/detailIngredient'
+
+const initialDetailIngredient = {
+  ingredient: {},
+}
+
+export const detailIngredientReducer = (state = initialDetailIngredient, action) => {
+  switch (action.type) {
+    case CLEAR_DETAIL_INGREDIENT: {
+      return {
+        ...state,
+        ingredient: {}
+      }
+    }
+    case SET_DETAIL_INGREDIENT: {
+      return {
+        ...state,
+        ingredient: action.ingredient,
+      }
+    }
+    default:
+      return state
+  }
+}
